refactor: extract Manifest() to manifest.go, fix templates and package comment
- Move Manifest() (~315 lines of data declarations) from google.go to
  manifest.go, keeping google.go focused on connector logic (~190 lines)
- Fix package comment to include "Sheets" alongside Gmail and Calendar
- Fix template duplication: locked-spreadsheet templates now use
  "SPREADSHEET_ID" placeholder, while "read any" template uses "*"
